First Log
Planning Phase and Technical Design Phases complete
Level Tiling complete
Basic side to side Movement (issue)

This is the first time I'm pushing. I don't know how to provide evidence that the planning is done at this time, but it is. This includes; hack'n'plan gannt chart, work breakdown UML class diagrams and a state diagram.

At this point the map is tiled and there is basic side to side movement on the player.
I'm having an issue where the seems to be invisible wall where there shouldn't be. At the moment I don't know if this is a collision issue or a script issue.

Currently on track with plan. Plan to implement Jumping and interaction with collectible. Will also try to solve the horizontal movement issue.